refactor: organize actions components
