+++ src/pages/adventure/Adventure.vue
@@ -1,6 +1,6 @@
 <script setup>
   import ArticleTemplate from '@/components/shared/templates/Article.vue'
-  import ActionsTemplate from '@/components/shared/templates/Actions.vue'
+  import ActionsTemplate from '@/components/shared/templates/actions/Actions.vue'
 
   defineOptions({
     name: 'BucketListTemplate',
