+++ src/pages/entertainment/Anime.vue
@@ -20,7 +20,7 @@
 
 <script setup>
   import ArticleTemplate from '@/components/shared/templates/Article.vue'
-  import ActionsTemplate from '@/components/shared/templates/Actions.vue'
+  import ActionsTemplate from '@/components/shared/templates/actions/Actions.vue'
   import animeList from './data.json'
 
   defineOptions({
