+++ src/pages/nutrition/ingredients/Ingredients.vue
@@ -1,7 +1,7 @@
 <script setup>
   import { ref, computed, watch } from 'vue'
   import ArticleTemplate from '@/components/shared/templates/Article.vue'
-  import ActionsTemplate from '@/components/shared/templates/Actions.vue'
+  import ActionsTemplate from '@/components/shared/templates/actions/Actions.vue'
   import ingredients from './ingredients.json'
 
   defineOptions({
