+++ tests/components/shared/templates/actions/ActionDeleteModal.test.js
@@ -1,11 +1,11 @@
 import { test, expect } from 'vitest'
 import { mount } from '@vue/test-utils'
-import DeleteModal from '@/components/shared/templates/actions/DeleteModal.vue'
+import ActionDeleteModal from '@/components/shared/templates/actions/ActionDeleteModal.vue'
 
 const action = { description: 'Test action' }
 
 test('clicking Cancel button or backdrop emits cancel', async () => {
-  const wrapper = mount(DeleteModal, { props: { show: true, loading: false, action } })
+  const wrapper = mount(ActionDeleteModal, { props: { show: true, loading: false, action } })
 
   await wrapper.find('button.btn-outline-secondary').trigger('click')
   expect(wrapper.emitted('cancel')).toHaveLength(1)
@@ -15,7 +15,7 @@ test('clicking Cancel button or backdrop emits cancel', async () => {
 })
 
 test('clicking Delete Action emits confirm and respects loading state', async () => {
-  const wrapper = mount(DeleteModal, { props: { show: true, loading: false, action } })
+  const wrapper = mount(ActionDeleteModal, { props: { show: true, loading: false, action } })
 
   let deleteBtn = wrapper.find('button.btn-danger')
   await deleteBtn.trigger('click')
